% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amscale.R
\name{amscale}
\alias{amscale}
\title{Aldrich-McKlevey Scaling}
\usage{
amscale(x, respindex = NULL)
}
\arguments{
\item{x}{A dataframe or matrix containing integer values of respondent placements of stimuli.}

\item{respindex}{An optional integer giving the column index of respondent self-placements.}
}
\value{
A list containing three objects.\preformatted{    stimuli contains a double vector of the scaled stimuli.

    respondent is a dataframe of the same length of the input containing respondent
    intercepts, weights; and if respindex was specified self-placements and ideal points.

    fit is a single value for AM's fit statistic for the model.
}
}
\description{
An implementation of Aldrich-McKlevey's (1977) scaling method for perceptual data.
}
