% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enp.R
\name{enp}
\alias{enp}
\title{Effective Number of Parties}
\usage{
enp(p, type = 3, S = NULL, range = F)
}
\arguments{
\item{p}{Numeric vector of party vote or seat shares. If \code{type} is 5 or 6, numeric vector of seat counts, see details.}

\item{type}{Single integer denoting which measure to use. See details.}

\item{S}{Single integer denoting the size of the legislature. Must be used for \code{type} 5 and 6, ignored otherwise. See details.}

\item{range}{Function returns minimum and maximum values along best estimate for \code{type} 5 and 6, ignored otherwise. See details.}
}
\value{
Numeric vector of length 1
}
\description{
\code{enp} implements several measures assessing the effective
number of parties or similar concepts. Defaults to Laakso and
Taagepera's (1979) Effective Number of Parties. See details for
more information.
}
\details{
Supplying the integer corresponding to a particular measure to the type
argument will mean that the function uses that measure. Defaults to Laakso
and Taagepera's (1979) Effective Number of Parties (\code{type = 3}).

\strong{Type 1}

Rae's (1968) fractionalization index. Ranges from 0 to 1. See
Wildgen (1971) for critique. Calculated as:
\deqn{1-\sum_i^M{p_i^2}}{1-sum(p^2)}

\strong{Type 2}

Hyperfractionalization index first used in Kesselman (1966),
further elucidated in Wildgen (1971). Calculated as:

\deqn{antilog(-\sum_i^M(p_ilog(p_i))}{antilog(-sum(p*log(p)))}

\strong{Type 3}

Laakso and Taagepera's (1979) Effective Number of Parties. Calculated as
\deqn{\frac{1}{\sum_i^M{p_i^2}}}{1/sum(p^2)}

\strong{Type 4}

Molinar's (1991) NP index. Calculated as:
\deqn{1 + N\frac{\left( \sum_i^M p^2\right) -p_1^2}{\sum p^2}}{1 + N*((sum(p^2)-p1^2)/sum(p^2))}

Where \eqn{N} is Laakso and Taagepera's (1979) Effective Number of Parties.

\strong{Type 5}

Taagepera's (1997) Effective Number of Parties for incomplete data.
Here, \code{p} should be a vector of seat counts instead of a vector of proportions.
It should contain only the seat counts for known parties (i.e. not decomposed
into other).

\code{S} must be specified and is the total number of seats.

This is based on Laakso and Taagepera's (1979) Effective
Number of Parties, using counts instead of proportions or percentages. This
is given by:

\deqn{\frac{S^2}{\sum_i^M{p_i^2}}}{S^2/sum(p^2)}

where \eqn{S} is the total number of seats. Taagepera shows that the
plausible range of the effective number of parties is:

\deqn{\frac{S^2}{R^2 + \sum_i^M{p_i^2}} < N < \frac{S^2}{R + \sum_i^M{p_i^2}}}{S^2/sum(R^2 + p^2) < N < S^2/sum(R + p^2)}

where \eqn{R} is the number of seats allocated to 'other' parties and \eqn{N}
is the effective number of parties. The best estimate is the mean of the two
extremes, and this is the estimate returned by 5. If \code{range = T}, these
are also returned.

\strong{Type 6}

Extension of \code{type = 5}. Calculations assume that no component
of \eqn{R} is larger than the smallest element of \eqn{p_i}{p}, \eqn{p_L}{pL}.
The plausible range used for calculating the best estimate becomes:

\deqn{\frac{S^2}{Rp_L + \sum_i^M{p_i^2}} < N < \frac{S^2}{R + \sum_i^M{p_i^2}}}{S^2/sum(R*pL + p^2) < N < S^2/sum(R + p^2)}

As before, if \code{range = T}, the minimum and maximum are also returned.

\strong{Type 7}

Taagepera's (1999) alternative NP index.
\deqn{1 + N - (\frac{N}{(\frac{1}{p_1})^2}}{1 + N - (N/(1/p1))^2}

where \eqn{N} is Laakso and Taagepera's (1979) Effective Number of Parties.

\strong{Type 8}

Golosov's alternative (2010) measure. Calculated as
\deqn{\sum_i^M(p_i/(p_i + p_1^2 - p_i^2))}{sum(p/(p + p1^2 - p^2))}
}
\references{
\insertRef{kesselman1966}{psmisc}

\insertRef{rae1968}{psmisc}

\insertRef{wildgen1971}{psmisc}

\insertRef{laakso1979}{psmisc}

\insertRef{molinar1991}{psmisc}

\insertRef{taagepera1997}{psmisc}

\insertRef{taagepera1999}{psmisc}

\insertRef{golosov2010}{psmisc}
}
