% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pascal.R
\name{pascal}
\alias{pascal}
\title{Pascal's Triangle}
\usage{
pascal(rownum, element = NULL)
}
\arguments{
\item{rownum}{A length one numeric vector containing an integer value denoting the row of Pascal's triangle to extract the desired element from.}

\item{element}{NULL or a length one numeric vector containing an integer value denoting the element within the specified row of Pascal's triangle to extract. Required if \code{rownum} > 1.}
}
\value{
A length one numeric vector containing the element from Pascal's triangle to be returned.
}
\description{
A function to extract a single element from Pascal's triangle. Note that Pascal's
triangle uses 0-based indexing, so the first row is row 0 and the first element in
a row is the 0th element.
}
\examples{
pascal(0)
pascal(2,2)
pascal(4,3)

}
