% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.amscale.R
\name{summary.amscale}
\alias{summary.amscale}
\title{Summarising Aldrich-McKelvey Models}
\usage{
\method{summary}{amscale}(object, digits = 3)
}
\arguments{
\item{object}{An object of class 'amscale'.}

\item{digits}{Number of digits for rounding stimuli locations.}
}
\value{
\code{summary.amscale} computes and prints to console the most salient information from the Aldrich-McKelvey scaling procedure implemented in \code{amscale}.
}
\description{
Summarising Aldrich-McKelvey Models
}
