% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amscale.R
\name{amscale}
\alias{amscale}
\title{Aldrich-McKlevey Scaling}
\usage{
amscale(x, resp = NULL, polarity = NULL)
}
\arguments{
\item{x}{A dataframe or matrix containing numeric values of respondent placements of stimuli, with stimuli on columns and respondents on rows.}

\item{resp}{A optional numeric vector containing respondent self-placements.}

\item{polarity}{An optional integer giving the column index of a stimulus that you wish to have a negative value. All stimuli and respondent self-placements will also be coded accordingly.}
}
\value{
A list containing five objects:\tabular{ll}{
\code{stimuli} \tab Double vector containing scaled stimuli estimates. \cr
\tab \cr
\code{respondent} \tab Dataframe of the same length as the input containing respondent intercepts, weights; and if resp was not NULL respondent ideal points. \cr
\tab \cr
\code{fit} \tab The value of the fit statistic for the model. \cr
\tab \cr
\code{nresp} \tab The number of respondents in the initial input. \cr
\tab \cr
\code{ccases} The number of respondents scaled in the model after filtering for missing data.
}
}
\description{
Performs Aldrich-McKlevey's (1977) scaling method for perceptual
data using the QR decomposition method (Swatton 2021) for calculation.
}
\references{
\insertRef{aldrich1977}{psmisc}

\insertRef{swatton2021}{psmisc}
}
