% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordTable.R
\name{ordTable}
\alias{ordTable}
\title{Title Table of Summary Statistics for Ordinal Data}
\usage{
ordTable(
  data,
  vars,
  compute = c("median", "agreement", "min", "25q", "75q", "max", "N"),
  values = NULL
)
}
\arguments{
\item{data}{Dataframe containing the desired variables.}

\item{vars}{Character or numeric vector containing integers naming or indexing the columns containing the ordinal variables you wish to summarise.}

\item{compute}{Character vector instructing ordTable which descriptive statistics to calculate}

\item{values}{NOT YET IMPLEMENTED}
}
\value{
A dataframe containing the median, Van der Ejik's agreement statistic, min, 25th quantile, 75th quantile, max, N.
}
\description{
Title Table of Summary Statistics for Ordinal Data
}
